feat: use Makefile to build backend
The idea behind the backend is to use typescript with minimum overhead, which means:
- do not involve lambda layers if possible
- bundle dependencies together and reduce file size
- local development and deploy should be able to use the same files without any modification

Native typescript support from SAM achieves the first two requirements, but fails to delivery the third one. In order for it to work, it needs a `package.json` file beside each lambda function, which is not the ideal local dev structure. And this behavior is not open to modification (SAM does not expose esbuild APIs).

However, SAM allows to use Make as build method with Makefile, which make customizing esbuild process possible.